Simplify `HasBindings` by removing `Maybe`.
It literally only applies to functions, as a general category of thing. Or at least, we can divide it on types in the IR, instead of writing it down here.

Removing it is a strict simplification, see delta, and makes future changes a bit easier potentially. It also changes "this only applies to functions" from a thing that happens to be true but isn't easy to see by reading the codebase, to something that is explicitly noted with a `!matches!(item, Item::Func)`.

PiperOrigin-RevId: 747938928
Change-Id: I6c09aeec61e0662bf852a292c3c7db33910df51a