Make `has_bindings` a query in `db`.
The direct reason to do this, even today, is that there are redundant calls, and it is not trivial in what it does. (It recurses over datatypes, and so on.)  But also, I am pretty interested in adding more redundant calls.

I am thinking of adding a `visibility` field into the `Yes` case, so that you might reasonably ask, when generating bindings, "are the bindings for this _public_" (or `pub(crate)`, or nonexistent). And so you might, when generating bindings for a function, add `pub(crate)` to it if its own `has_visibility` returns `pub(crate)`. Or, when iterating over all items, we might add `forbid(private_interfaces)`, but only if bindings for any thing are `pub(crate)`. With these kinds of redundant calls, it makes even more sense to make `has_bindings` cheap to call.

But I'm not certain exactly how that will play out. Putting that aside, I think this makes sense to do today. I seem to recall there being a reason it wasn't possible before (maybe code organization?) but it's possible now.

PiperOrigin-RevId: 748067961
Change-Id: Ic6a8c990e8170827f80b00138039719d20953d89