Allow returning non-Rust-movable types by value in `:wrapper` mode (though not yet accepting parameters by value in any useful way)
I don't think we need to wait for move constructors to get bindings to submit this. :)

This also sets up the stage for how we might do `pub(crate)` wrapper mode stuff more generally -- this CL was basically me dipping my toes in the water to explore it, and a followup CL will dive into it and make entire types `pub(crate)` by nature.

PiperOrigin-RevId: 750306389
Change-Id: Ie22ddc3877a7d60a6212107b0dd71df797457953