Add RsTypeKind::unalias()
There are many places where an RsTypeKind variant is checked directly, when it should also be recursively looking through type aliases. This adds an `.unalias()` method that takes an `&RsTypeKind` and strips away all layers of `TypeAlias` to reveal the innermost `underlying_type`. I added it everywhere I think is appropriate.

This was causing me issues when I was trying to add support for composable bridging for std::string, because it's aliased so much in C++.

PiperOrigin-RevId: 750350687
Change-Id: I3cfcca7e9cb0237e53655e6301068615ba97b913