Accept `Rc<Foo>` in all `generate_foo` functions.
The status quo was a little confusing/annoying: the ones which are database functions took an `Rc`, the ones which weren't didn't. But this has the downstream consequences that when you do need an `Rc` after all (to generate an item), you do something silly like `Rc::new(x.clone())`, ew! There was one instance of this need already, and I was going to add more as part of needing to call `has_bindings()` from the implementation of `generate_foo` (to decide whether to return a token stream with `pub` or with `pub(crate)`.

This CL maybe makes things slightly less efficient, but in exchange we get consistency, which is nice, and it lets us more easily create an `Item` which becomes more useful later.

PiperOrigin-RevId: 750372228
Change-Id: Ieaa9ed5e357b27bdd0769a28ca859415116162c0