Document all `query_group!{}` functions.
Where the implementation function was documented, I moved the comment to the `db.rs` trait method, with at most minor wording touch-ups. Where it was undocumented, I wrote some minimal doc comment myself.

I also, across the board, added cross-references to the implementation for ease of navigation, though they won't be good as github links.

A quick note on macros: I tried to allow this by some fancy macro magic first, but it was way too annoying. The easiest fix is to swap the order of the `#[input]` etc. This resolves the parsing ambiguity, because Rust will know exactly which branch to enter very early on.

An alternative would have been to just use regular comments, but regular comments don't show up in IDE integration, while doc comments do.

Note that Rust-Analyzer will only show the comments on the function you call! So the right location for detailed comments is on the trait methods,
not the implementation. In fact, the best comment on the implementation is one that says, essentailly, to call the `db` method instead.

PiperOrigin-RevId: 750678279
Change-Id: I564b93bfe5212ec015b5d94e3ac70095dd700529