Generalize the `Item::TypeAlias` item check so that it can work for other types.
I'm going to be adding more branches here in a followup CL where I introduce types that are `pub(crate)`, and want it so that visibility of a type matches visibility of the RsTypeKind for that type, and `from_item_raw` is the glue that can patch these together.

Demoing it with just `TypeAlias` makes this pretty easy!

PiperOrigin-RevId: 751179670
Change-Id: I3747ce6b928990e82c2fb2308bfb36a5298ead38