Move failable logic for generate_enum directly into `has_bindings`.
This replaces a direct call to `generate_enum` **from** `has_bindings`. That approach doesn't work for structs, for example (see TODO!), where a struct can depend on itself in a cycle: `struct S {S* x;};`. If we copied the `enum` approach there, it would cause `generate_record(s)` to call `has_bindings(s)`. So the "generalized" fix is to move failable things into `has_bindings` where everyone can see them, without needing to call` generate_item`, which we might be in the middle of doing already.

Also, I want to actually directly call `has_bindings` within `generate_enum` etc. in order to determine the visibility of the item we're about to generate. :)

PiperOrigin-RevId: 751186613
Change-Id: Iabae143d5e5a3a72676b2035e89bb0aac4d1954b