First draft: `ffi_11` module, for 1:1 FFI types.
The idea here is to provide a substitute `ffi` module, which is "high fidelity: it maintains a 1:1 relationship between `ffi_11` types in Rust, and fundamental types in C++, and ensures that the mapping between fundamental types and sized types is the same in both languages. (E.g. `int64_t` is an alias for `long` if, and only if, `c_long` is an alias for `i64`.)

This CL is woefully incomplete. There is a lot more that needs working on, hopefully fully characterized in the module doc-comment. But in addition to what the doc comment says:

* This needs to be released with the other support libraries if we ever generate code using it. (For example, nullptr_t!)
* Eventually, this should be released externally -- I have the idea that this really could be pushed to crates.io and used by other projects.

### On naming

I'm a bit attached to `ffi11` or `ffi_11`, because it's a pun, meaning three things at once:

* `ffi`, but one-to-one (and onto).
* `ffi`, but it also adds C++11 (and above) support. Generally, anything C++11-aware is more modern – this is also reminiscent of, for instance, `pybind11`.
* And, of course… https://www.youtube.com/watch?v=4xgx4k83zzc

[]

Example snippets to help with my case:

*  `ffi11::c_uchar as ffi11::c_schar`
*  `ffi_11::c_uchar as ffi_11::c_schar`

PiperOrigin-RevId: 751596246
Change-Id: Ia41601a6aaf8430d84fd5717679aa3de17c241b1