Diagnose nonnull pointer fields that are nullable at method exit.
This was a gap in our diagnosis of moved-from nonnull smart pointers: We warned
if they were accessed within the same function that left them in the moved-from
state, but we didn't warn if a nonnull smart pointer field was left in the
moved-from state when a method exited.

In addition, this now also warns if a nonnull raw pointer field is null at
method exit.

PiperOrigin-RevId: 752391452
Change-Id: I88c7f9e925063c3722558d6ede49b3100d6706c1