Enable all features known by the binary in integration tests / goldens.
This avoids some of the "wait for a binary release" annoyance during development. If you add a new feature flag, you can start using it immediately in tests, including integration tests, without waiting for a rollout of the feature so that it can be added to the `:experimental` feature list.

It also, as a side effect, dramatically expands the features list -- only partially ameliorated by https://github.com/google/crubit/commit/fb3fa5a4255ff4d9a306733dcff439891b3000b3. I considered if maybe we should check if all features are enabled and, if so, stringify to just `all`, but then there's some ambiguity for what that means (depends on which version of the binary). I'm not sure what call to make there, and want to leave it to a followup CL.

## Release-safety

I actually mailed this out and then retracted it, and am now re-mailing it. There's a very significant question here: is this safe, or does this make binary releases too risky?

The reason we have feature flags is exactly to decouple changes to Crubit's behavior from the binary release process. If we make a backwards-incompatible change, it gets deployed via a flag flip.

The problem is, `:experimental` already contains ~everything, or is intended to! It is the default living space for all new features, which get migrated to :supported via a flag flip.

So while it's not safe, it's not a very big regression: generally speaking, the only things we put in a feature flag are those we are migrating from `:experimental` into `:supported`. And having the flag list for `:experimental` be explicit forces this to take a minimum of two binary releases to accomplish, instead of one.

It's true that unsafe_types was removed from `experimental`, but that was an accident of https://github.com/google/crubit/commit/d4c81be533618a0b6eac8025c49fb06b9c51ba51, and it's going to be immediately re-added once that change is released (which was the point of the change).

(You might ask: why not change it so that the `experimental` string expands the same way as `all`? Well, for one thing, I'd need to wait for another binary release! Beyond that, it might be a good idea.)

Note that if we add a separate feature set for "broken" features, we would want to exclude them from :experimental.

## Golden diffs

Most of the golden diffs are for the unsafe feature. I manually checked that each diff was OK and still reflected intent of the test:

* bitfields: fixing the golden would actually defeat the test: we want the struct to have unknown fields, because that's the point of the regression test.
* typedefs: expected, unions are unsafe now
* types: expected, the struct has pointer fields and so is unsafe
* unions: expected, unions are unsafe now
* unsafe_types_transitive: the test is actually fixed now :)
* unsupported: expected, unsupported fields are assumed unsafe now
* user_of_imported_type: expected, actually useful to ensure that unsafe crosses crate boundaries like this

PiperOrigin-RevId: 754752151
Change-Id: I9850e49c06a676b7291f68cf951a7f02d34af0e7