Diagnose nonnull pointer fields that are nullable at method exit.
This was a gap in our diagnosis of moved-from nonnull smart pointers: We warned
if they were accessed within the same function that left them in the moved-from
state, but we didn't warn if a nonnull smart pointer field was left in the
moved-from state when a method exited.

In addition, this now also warns if a nonnull raw pointer field is null at
method exit.

PiperOrigin-RevId: 754840691
Change-Id: Id92f74f137008abc4d2d2f0b408900989295dddd