Disable cc_bindings_from_rs goldens in release directory.
The problem is that since we generate a path-based include guard, when the cc_std directory is copied over, the guard value changes, causing a diff / failing test.

This isn't a super extensible approach (how would it work in other repos?) but for now it does work.

Alternatives:

* Move the test outside of cc_std, so that it doesn't get copied over. But I think without some kind of presubmit, we'll just forget and do it again.

* Check instead for //third_party/crubit. But that seems surprising in its own way (now the golden test infra is tied to this directory specifically, and we can't refactor things outside of it without surprising ourselves with tests not running).

* Do some kind of build transform when we copy over. That's probably better, but I didn't know how to do it at a glance.

This seemed reasonably straightforward and durable. It won't be hard to replace it with a better approach later.

PiperOrigin-RevId: 754844032
Change-Id: I2ef6b58bd34a63d6a0a8689193f96a07b726ca83