Basic update of the cargo workspace.
This wasn't great -- I think the intent had been to convert this to an aspect so that these can be generated on demand instead of being manually listed out one by one.

Also, this has broken more recently by having two crates named `database` (etc.) in the same workspace. I'm not fully sure how to fix that, but anyway, that's a problem for a future CL.

My hope had been to try to get `ffi_11` something I can just run `cargo <command>` on, but that needs the entire workspace to be green!

PiperOrigin-RevId: 755803949
Change-Id: I1076a2a87efe79200f3304509da14ab5141efcc6