Add support for pub(crate) types in `:wrapper`, starting with forward declarations.
The central change here is as follows:

* Directionally, we're going to change rs_type_kind to _always_ return something, even if that "something" is only used in `:wrapper` or `:experimental`.
* Filtering is done after-the-fact by  `required_crubit_features` / `has_bindings`, instead of ahead-of-the-fact by returning `Err` from `rs_type_kind`.
  * This has a side-benefit of allowing better and more complete error reporting, since we can get further into bindings generation with types that won't work by the end of it!
* `generate_foo` should call `has_bindings` on the RsTypeKind for `foo` to see what visibility to generate. One of them has to call the other to keep logic in sync, and I picked this direction because we actually _already_ need to call `has_bindings` from `generate_item` when constructing cyclic records. (The purpose of `has_bindings` is to allow this sort of shared logic with cycle breaking.)

This is not _obviously_ the right move. The other logical change we could make is to make `rs_type_kind()` take an additional parameter, an `Option<BazelTarget>` for the crate that the type is being used in. However, that involves changing all callers, in a way which is actually mostly incidental complexity: the core logic needs to go in `required_crubit_features` / `has_bindings` *anyway*, because _that_ is where we do feature flag based filtering, and we are doing this in a feature-flag-dependent way.

In addition, if we _do_ want to change `rs_type_kind()` in that way, that change can be built on top of this one -- effectively, by adding an extra call to `type_visibility_restriction` in `rs_type_kind()`. So this is a smaller CL for not doing that.

I'm not sure that there's a substantial risk to this approach / a substantial reason to move the logic into `rs_type_kind()`, but we might discover more over time. We don't need to make the decision now, IMO, and can opt for the smaller/easier CL.

(My first attempt at a CL was to put the logic in `rs_type_kind()`, but it in fact was actually quite a bit more work!)

PiperOrigin-RevId: 755806267
Change-Id: Id4b735942be37e727598ff28c7bd058f39b094c1