Make class templates available as `pub(crate)` types in `:wrapper` mode.
This is a bit of a test of the overall approach: we got it working for forward declarations, how hard is it to get it working for class templates? The prediction was: oh, it's identical, except that we need to special-case `string_view`. And that came true to the letter: we just needed to refactor a little to make the `string_view` check visible, and call that from the same place we do the visibility restriction logic for forward declarations.

PiperOrigin-RevId: 755831899
Change-Id: Icb7e4dc6cd81d7e51288472c6abdc763c7c45259