Move most of the `ffi_11` design doc to a README.md.
The design doc explained the problem and tried to justify the solution -- honestly, not a bad thing to have as a README.md, in my opinion!

So the idea is, the README.md can introduce why this crate exists, and the API docs have concrete details about the exact invariants and functionality.

Not included from the design doc:

* The API docs that became part of the crate doc-comment.
* The "alternatives considered" section, and open questions, and so on. (These might deserve a home, but to be honest, the README.md is already pretty big!)
* The detailed breakdown of what it might look like on different platforms. To be honest, that still should have a home. The fact that WASM's integer type categorization is so cool deserves to be a published fact. But it didn't seem to read well as a README per se.

PiperOrigin-RevId: 758071955
Change-Id: Id77559f981ccfb5058b95ac2ec2f390f6d892cf7