Allow Status to be remotely bridged
Abseil is very hard to add deps to because it's depended on by many other libraries, which is problematic for bridging because we need to add `:bridge_cpp` as a dep to `:status`. This CL introduces a bandaid fix to remotely hardcode `absl::Status` as a bridge type, similar to `std` types like `optional`. This CL doesn't do anything yet because `:status` isn't annotated to support Crubit yet.

PiperOrigin-RevId: 766289492
Change-Id: Icb8a727962eaf41de28ea8e0792877b8c931a471