Fix `primitive_types_test` on macOS.
This was comparing `isize` and `usize` against `int64_t` and `uint64_t`,
respectively. On macOS, however, these are different types: We map `isize` to
`intptr_t`, which on macOS is an alias for `long`, while `int64_t` is an alias
for `long long`. (These are both 64-bit signed integers, but different types.)

Using `intptr_t` and `uintptr_t` resolves this issue, while still testing what
we actually want to test. In addition, this means the test will pass once we
add support for 32-bit platforms.

PiperOrigin-RevId: 766323755
Change-Id: Ib5bf414029dee1daade7eb1aac75589551505e04