Fix broken `PartialEq` implementation for `arc_anyhow::Error`.
I'm pretty sure I did this, and meant to do pointer comparison, in the sense of checking that two Arcs refer to the same thing. Instead, apparently, I was comparing the address _of_ the Arc. Uh.

This didn't matter at all before, because when we actually needed this with Salsa, the address was stable enough. Worst case (like with hash table resizing) it would lead to an unnecessary recomputation. But I want to insert these into a hash map in a followup CL, and there it gets broken.

PiperOrigin-RevId: 766719307
Change-Id: I49425eeb8c374fce12c8444d487bdf0ed1afbb14