Fix inherited inline methods. (And other inherited methods that require a thunk.)
Once again proof that spooky action at a distance is a source of pain. Here, one part of the code thought that it was going to reuse the existing thunks, and another thought it was going to be creating new thunks. I picked a winner -- it's strictly less code generation if we reuse the existing thunk, and I think that was the intention. It seems to work, so, yay!

PiperOrigin-RevId: 767016034
Change-Id: I6c8d3ff7ec2c267c4db022a6d6b0782cef146c3c