Start refactoring ApiSnippets.main_api
This CL does nothing but change ApiSnippets to store a `Vec<MainApi>` instead of `TokenStream` for it's `main_api` field. Followup changes will be made to make the variants on `MainApi` more detailed.

This is incremental progress towards having Crubit spit out a fully structured but not yet stringified API, with the idea that we can then choose how to name things in generated code since we have full view over the entire generated API.

PiperOrigin-RevId: 767251495
Change-Id: I2ecc5046bdbdf0634cf8d46b842106ec4e496b23