Allow spelling `impl Ctor<Output=[forward declared type]>`, even if its implementation is just a panic.
This is the easiest way to make comprehensive fallbacks work (when implemented): if a type is stubbed out and replaced with an incomplete type, for error purposes... that's still OK! You'll get a compilation error if you try to call it, but not a type error to define this error case.

In the absence of comprehensive fallbacks, I test this using a similar example for slice.

PiperOrigin-RevId: 769505793
Change-Id: I4c86749ea84aac35b4c3cd612efae41bbdc291e3