Non-comprehensive fallbacks.
This CL introduces the notion of an error type, which can lead to immediate failures (in most build modes) or else can be replaced with a forward-declaration.

## The completeness (or lack thereof)

This CL only falls back in the following circumstances:

1. The type is an error for non-visibility-related reasons.
2. the type is used by pointer or by reference.

### (1) visibility

Regarding (1): visibility-related fallbacks require us to know what crate we're in. I've been hoping to avoid this, as it's a major refactor, and I got this far without it... but it really does seem like there won't be a way around it.

I expect a future CL to work like follows:

```rust
/// the old rs_type_kind function,
/// returns the "globally visible" version of a type.
/// Never called except by the new rs_type_kind function.
pub fn rs_type_kind_impl(cc_type)

/// rs_type_kind for this target.
/// if a type is not visible to this target, replaces
/// with a fallback type.
///
/// calls into rs_type_kind_impl for shared caching when possible, idk.
pub fn rs_type_kind(cc_type, target: BazelLabel)
```

### (2) by-value

A future CL will need to change the bindings for functions (and only functions!) so that return types get heap-allocated, and parameter types are transformed to be by pointer.

## The forward declarations

I had worried we'd need to collect the error types, add a single forward-declaration, and [...]. This is pretty hard.

Instead, in this CL, we rely on structural typing: basically, `Incomplete<symbol!("..."), ...>` will always be the same type, no matter how many times you say it, and so we do not need a list in advance of forward-declared types.

This is WAY WAY WAY WAY WAY less work than trying to thread through all the types we depend on, collect them, and generate a forward_declare!() invocation for each one. Trust me! I tried it! Two different ways!

A final note -- we can helpfully use `()` as the crate identifier, because the type is private and so collisions don't matter. This is more straightforward than generating the crate-local type. :-*

PiperOrigin-RevId: 769527799
Change-Id: Ic059ba727585a4bca8fb3fd6913ae71bf409f927