Comprehensive fallbacks: fall back for bridge types.
This isn't as good as really exposing the original type, but it's something. I'm most interested this as a preparatory step for doing this for protos, which are also some flavor of bridge type, but for which the original type won't be available since we don't run Crubit at all on the generated C++ headers.

PiperOrigin-RevId: 769562292
Change-Id: Idab7b3ebd73e0b472529258c09658a86cca744c0