Use `unsized_const_params` in `forward_declare` after all.
`forward_declare` wants to craft structural type equality -- if two strings have the same value, then two types are equal.

Before this CL, `symbol!("foo")` accomplished this by expanding to a tuple, `(C<'f'>, C<'o'>, C<'o'>)`. This works, but it's patently unreadable. It's worse than I thought, though. As soon as I started using `forward_declare` "for real", I discovered that in type errors, the name can be so long that rustc refuses to actually print it entirely:

```
   = note: expected raw pointer `*const Incomplete<Symbol<...>, ()>`
              found raw pointer `*const Incomplete<Symbol<...>, _>`
note: function defined here
  --> .../blahblah.rs:17:26
   = note: the full name for the type has been written to '.../blahblah_test.long-type-7711691867200381589.txt'
   = note: consider using `--verbose` to print the full type name to the console
```

Now, you might imagine you could just read the file, and it's NBD, but in my experience with our build system, the path is not actually loadable. It was some file in an intermediate step in the build and is not an exported artifact, or something. Or maybe I'm misdiagnosing it, but at any rate, if I copy-paste the path I don't see it! And if someone sends me a link to the broken build, I have to guess what might be there! The error is awful.

Improving the error and actually showing the full type names requires the above unstable feature, and when I last checked, Augie was kinda sorta -0 on this one. Actually, it's an especially concerning one, given that rust warns against it with `incomplete_features`.

So, to make this easier, I not only keep the current version around under a flag, I make sure it's tested. So if for whatever reason the unstable feature ends up being a bad idea, or removed, or breaking in some release, we can always fall back to the old approach. The only cost will be worse error messages.

---

I couldn't unit test the type error, but here's a manual test of what it looks like with the unstable/stable feature:

Before this CL / in the stable test:

```
error[E0308]: mismatched types
  --> third_party/crubit/support/forward_declare_stable_test.rs:38:9
   |
37 |     fn _is_same(x: *const Type1) -> *const Type2 {
   |                                     ------------ expected `*const Incomplete<Symbol<...>, ...>` because of return type
38 |         x
   |         ^ expected a tuple with 27 elements, found one with 23 elements
   |
   = note: expected raw pointer `*const Incomplete<Symbol<...>, ...>`
              found raw pointer `*const Incomplete<Symbol<...>, ...>`
   = note: the full name for the type has been written to '.../forward_declare_stable_test.long-type-2806037806920280064.txt'
   = note: consider using `--verbose` to print the full type name to the console
```

And by the way, the written file does not exist:

```rust
$ cat [...]/forward_declare_stable_test.long-type-2806037806920280064.txt
cat: [...]/forward_declare_stable_test.long-type-2806037806920280064.txt: No such file or directory
```

After this CL:

```
error[E0308]: mismatched types
  --> third_party/crubit/support/forward_declare_unstable_test.rs:34:9
   |
33 |     fn _is_same(x: *const Type1) -> *const Type2 {
   |                                     ------------ expected `*const Incomplete<Symbol<"This type is something else">, _forward_declare_Type2>` because of return type
34 |         x
   |         ^ expected `"This type is something else"`, found `"hello::world::FooBarBaz"`
   |
   = note: expected raw pointer `*const Incomplete<Symbol<"This type is something else">, _forward_declare_Type2>`
              found raw pointer `*const Incomplete<Symbol<"hello::world::FooBarBaz">, _forward_declare_Type1>`
```

Perfect!

PiperOrigin-RevId: 769986023
Change-Id: Ic96b082759b0337316b787dcae601c47f7755f9b