Warn on more cases of nonnull fields that are initialized to nullptr
Track the nullness of nullptr through:
- value initialization through ImplicitValueInitExpr (we already handle CXXScalarValueInitExpr), but this handles the case through initializer lists
- `= nullptr;` through CXXDefaultInitExpr, in initializer lists

Otherwise, we get the nullability from the type which would be incorrectly Nonnull due to the annotation on the field.

We may want to enhance the diagnostic to indicate *which* field has the violation. Right now it just highlights a spot in the initializer list, which could be the end of the list.

PiperOrigin-RevId: 771147877
Change-Id: If61f7668eadf9b9bbda70e7d7f89c7e1505c0fef