Remove "type map override" feature gate
This feature gate breaks round-tripping of types defined in Rust
which appear in C++ APIs. The feature gate itself is not serving a strong purpose,
as manually defining a C++ type that is equivalent to a Rust type is already an
expert feature and requires importing Crubit-internal support macros. We don't
need to create extra friction for users of this feature.

PiperOrigin-RevId: 773892075
Change-Id: Ib66f9de21153c78c6be580026cab412b271fb86f