Fix forward declarations by not allowing them to combine with aliases.
This is a bit of a messy fix. We're acknowledging that the view of a type might be different from different translation units -- but claiming that the only place that this causes issues is when the different translation units _need_ to know how the other ones see it. This only happens with type aliases, where library A might use a type alias, and library B might use `A::TheTypeAlias`, but expects to be able to know properties of the type behind that type alias. It only knows what it sees, not what library `A` saw, and the result can be buggy. So, by banning this, we fix forward declarations, in that the existence of forward declarations should no longer cause issues. But it doesn't fix the fundamental problem.

Does this happen anywhere else?

* Functions: no. We never generate code which calls functions from another library.
* Records: no. The properties of a class don't change if we complete a type inside one of its fields (because it's behind a pointer), and we don't generate code which relies on anything in other libraries except the abstract properties of the classes.
* IncompleteRecord: not directly a concern.
* Global variables: same rough structure as functions, no.
* Enums, comments, namespaces, UseMod: definitely not, by structure.
* UnsupportedItem: at least not inherently, but maybe as a corner case of one of the above.

Type aliases do seem to be the only place where we use something from another library, need to know its properties, and those properties can depend on the completeness of a type.

PiperOrigin-RevId: 774798490
Change-Id: Ib3f88e35af88a58453e58c8d2b9bb269e159ebc6