Add an error output type to `Ctor`.
This is likely to be necessary for any broadly-usable variant of `Ctor`, because, well, initialization can fail, and not everyone is OK using panics. For example:

* The linux kernel's `pin-init` crate has an error parameter, and may be moving to an associated type, see https://hackmd.io/@aliceryhl/BJutRcPblx
* In environments where exceptions are enabled, we may want to have non-noexcept constructors return a Result, in case they throw. At the least, some users may prefer this, judging by conversations during RustWeek. So for Crubit to work for these environments, it needs to be able to do something other than panic.

This does not make the error type fully work. Almost all of the supporting machinery is hardcoded to only work with `Infallible` error type. But it's the start / it shows that it's not a supremely breaking change, and it's better to do this now (before `Ctor` is released) if we want to do it at all.

#### Verbosity

The result is pretty dang verbose.

What used to be `impl Ctor<Output=T>` becomes `impl Ctor<Output=T, Error=Infallible>`. Except of course everything is fully qualified, so it's more like `impl ::ctor::Ctor<Output=::somecrate::path::to::T, Error=::ctor::Infallible>`.

Some of this will get fixed by time: one day, it will be `impl ::ctor::Ctor<Output=::somecrate::path::to::T, Error=!>`.

If that's still not enough, then there's a handful of options.

One alternative would be a `FallibleCtor` trait, and a `Ctor` alias for infallible ones.

Something like:

```rust
pub trait Ctor : FallibleCtor<
    Output = <Self as Ctor>::Output,
    Error = Infallible> {
  type Output : ?Sized;
}
impl<C: FallibleCtor<Error=Infallible>> Ctor for C {
  type Output = <C as FallibleCtor>::Output;
}
```

This would make interfaces simpler (one can accept/return just an
`impl Ctor<Output=T>`), but at the cost of having two traits.

Another option would be to use trait aliases (https://rust-lang.github.io/rfcs/1733-trait-alias.html) for real. Not just to implement a `FallibleCtor`/`Ctor` split, but we could do even better than that: we could give each generated type a canonical trait alias for how it's typically constructed without errors, and then it would be:

```
impl ::somecrate::ctor::path::to::T
```

Where the generated `ctor` submodule looks like:

```rust
pub mod ctor {
  pub mod path {
    pub mod to {
      pub trait T = ::ctor::Ctor<
          Output = crate::path::to::T,
          Error = ::ctor::Infallible>;
    }
  }
}
```

So I think this may be workable on the whole, but I didn't want to explicitly pick a winner early on.

PiperOrigin-RevId: 774891694
Change-Id: Ic0b5e2ad21526372c83ef69a9a2db996fa434bb0