Generate bindings for types with lifetime parameters
The ban on types with lifetime parameters was initially introduced as a
conservative measure meant to protect against potential temporal and aliasing
violations. However, in practice it has motivated users to create fake
lifetime-less structures using raw pointers, resulting in increasingly unsafe
Rust APIs.

This is an initial first pass which supports basic bindings for these types.
Many uses are still limited by the "multiple uses of a lifetime parameter"
restriction which will be relaxed in a followup change. In order to make that
change, restrictions will have to be made in order to avoid allowing C++
references to be bound to Rust references with lifetimes that are nontrivial
(e.g. which outlive the function or must match another named input), as such
references are error-prone in C++.

PiperOrigin-RevId: 775311232
Change-Id: I5887b1f756ee3a1430d492da883c5f36a1a87f10