Make true forward declarations public in wrapper libraries.
This is a hacky work around for an unfortunately difficult problem: if they are pub(crate), then Crubit won't know, when looking at a _different_ library, that it needs to write fallback types for it. Instead it just fails. See the newly improved test coverage, ish. (That turned out messier than I hoped.)

After this CL, comprehensive fallbacks for pointers ~works, modulo much smaller gaps, presumably. I'd move my focus onto supporting fallbacks by value.

---

The reasons to make them pub(crate) are:

1. Originally, forward declarations were broken. See b/318690257.

2. visibility restriction -- until forward declarations are generally available, it would be nice to restrict their visibility, so that changing their APIs only breaks O(1) targets.

I fixed the bug in (1) exactly so that it would stop being a reason against this CL. As for #2, going for a cost/benefit calculation, we can just expose them publicly and the blast radius will still not be gigantic. We want to release them as part of `:supported` anyway, so this is just being a little bit aggressive.

The refactor necessary to make these remain `pub(crate)` would have been quite large (it requires every caller to `rs_type_kind` specifying which target they are currently in), and it would only be necessary just to support this one corner case. I think the tradeoff here is worthwhile, let's just do it.

PiperOrigin-RevId: 776235188
Change-Id: I9325bfecb5de0226c9fc5e5e52975864447a4fda