Remove feature gates from drop impl bindings
https://github.com/google/crubit/commit/dc5ea8a91ae9bc7a4a5df0faac5a645621f77376 unblocked generation of types with lifetimes.
However, prior to this CL we would still generate an error when attempting
to create types with both lifetimes and drop glue since the signature of
`Drop` contained multiple references, one of which was mutable, and we've
feature-gated the ability to access such functions under
PossibleMutableAliasing.

This CL disables all feature gating based on the generated API for the Drop
trait, as the resulting function isn't regularly callable by end-users.

PiperOrigin-RevId: 777798426
Change-Id: I344c25e09c43959d8ed1e26e7337d0a40a06a7c7