Make some database functions not memoized
I want to make some `generate_{record,enum,item}` functions write to an out parameter, instead of spitting out a brand new `ApiSnippets`. These functions are only called once with each input anyways, so it makes sense to make them non memoized so they can take an out parameter instead. `generate_function` theoretically should be here too, but it's used by `overloaded_funcs` right now so memoization is important until it can be replaced in one atomic change.

The reason we want to keep using them as db function though is so we can access them through function pointers, keeping compilation times to a minimum. To do this, I made a vtable `#[input]` and a `#[provided]` function for each one to keep good ergonomics. The only change for `query_group!{}` this required was making provided functions be implemented in the database impl of the trait, instead of in the trait itself because of Sized + dyn compatibility reasons.

PiperOrigin-RevId: 777804792
Change-Id: I33a070917e1fa89ea9245eacc8f10814b5ea1485