Fix make_rust_abi_path function
This is the third revision of this function- let's walk through each one.

The first revision of this function is in https://github.com/google/crubit/commit/37e47a3dbda941eced4e8a5731fe0dc1ba984309. If the path started with "::", it was left unchanged. If it didn't, but was in the current target, it was prepended with `crate::`. Otherwise, it was prepended with `target::`, where `target` was the name of the target. The problem here being that it should have generated `::target::` in the third scenario, to make it fully qualified.

The second revision in https://github.com/google/crubit/commit/8bc582d303ee08d2a66580f5bd35836ad435cad7 fixed that by prefixing with `::target::`, but made another mistake: it would _always_ prepend either `crate::` or `::target::`, even if the path already started with `::` (meaning it was already fully qualified). This was fine for all existing cases, but new cases like my StatusOr CL which uses `::status::absl::StatusOrAbi` were getting converted to `::statusor::status::absl::StatusOrAbi`, which is incorrect.

This CL is the third revision of this function (and hopefully the last), and respects the leading `::` by not changing the path at all if it's present, while still prepending `::target::` when we're not the target.

PiperOrigin-RevId: 780134120
Change-Id: Icc77f6dc8c4825371fe479ede1715f152319d722