If a single function in an overload set is "most canonical", use bindings for it.
The idea being that it's better to produce bindings for something than not to produce bindings for anything, and here we have a pretty clear signal on how to resolve the ambiguity of an overload set. If all but one function is deprecated, well, they're usually just weird ways of spelling the same thing, and we should at least export the current best way!

I could not come up with a structure I found especially happy, and my sleep-deprived butt spent way too much time on this CL, but something roughly like this seems right.

PiperOrigin-RevId: 780165008
Change-Id: Ia585b4d82b05798402c9f742c8b1fa914e109dfe