Only generate template bindings if their type arguments have bindings.
Theoretically speaking we should be able to get away with more than this. If a type argument does not have bindings, we can still generate bindings for everything that doesn't _use_ that type argument. But that actually includes "everything that spells out the name of the type", including the static assertions. So it's some work to unwind that, and may be easier to only generate bindings for types we can name.

Instead, I would suggest expanding the set of bindings by expanding the set of types we can name. For example, by understanding "canonical" names that might not be the definition in the `private:` section, as suggested in b/430028665#comment5

(And maybe the end result is that we always try to be able to name a type, in the worst case by some kind of decltype() call, getting us back to being able to generate bindings for ~everything.)

PiperOrigin-RevId: 782054836
Change-Id: If016e8c54e9dc8e43c1d6627ff80fa0d6b1c4af4