Proper name conflict detection for nested items
This works by modifying `has_bindings.rs`. When there's an item that should appear in a nested module, it first goes to the [Rust namespace](https://doc.rust-lang.org/reference/names/namespaces.html) that the parent record lives in, and performs name resolution on all the items in that namespaces. The name resolution logic takes place in `has_bindings::resolve_type_names`, and essentially prioritizes existing items over the generated nested module. Then, it looks up the name the nested items module would take on- if the name uniquely resolved to that nested module, great! Otherwise, that nested module cannot be generated, so the item that should live in it shouldn't be generated either.

PiperOrigin-RevId: 783059578
Change-Id: I403dd486c6c42e61b3dcf1fa69e1f01d4ab25954